extend tests to use item with a range key.
This exposes a bug (previously filed as #90) which occurs when items with a range key are read
from the DDB event stream, an md5 hash of the key is computed and the item written to S3.

The issue is that the DDB event stream handler does not (and should not) do a 'describe_table' to
know which key is the HASH and which is the RANGE and therefor simply generates the md5 hash of
the item keys in whatever order they happen to appear in the stream event.
The s3-backfill util does do a 'describe_table' and does order the keys by declaration
order which DDB requires to be HASH first, RANGE second.

The different ordering of the item keys will produce a distinct md5 hash value and different S3
path/keys will result in some items appearing twice in S3 effectively corrupting the incremental
backups since two valid versions will be present at the same time.