Add trailing line breaks to incremental backups
Up to now incremental backups didn't contain a trailing line break at
the end of the file. When aggregating incremental records into a
snapshot `dynamodb-replicator` would add them.
This commit changes that to include trailing line breaks in the
incremental backups and only add them during snapshotting when they're
not present.
The motivation for this change is that it makes it significantly easier
to process incremental backups with tools such as the `awscli` if they
already contain the trailing line break.